First half of persistent scoreboard
When developing FFA it's frustrating to players if the server
restarts. Their scores and upgrades are lost and many take that
as a stopping point for gameplay.

This is a feature that will ensure if the server restarts that
player data is stored for recovery. Haven't yet decided on the
exact behavior. Some options:

- a /server command that triggers capture of stats
- continuous periodic stats capture
- only capturing data on sigint, sighup, or some other signal

That's about as far as this is for now.