Reading committed logs at startup.
This commit adds the ability to read from the recovery file on app
startup.

- recovery file is read when application starts up
- recovery file contains just enough info to bootstrap initial state
- recovery file is written every minute. this may be a problem
- on shutdown, the recovery saveAll function is called via an event
- when the saveall function exits, it emits an event that exits the app

A couple of things left - right now the default behavior is to read from
the recovery file. I may add a server command to trigger a "fresh"
start, prompting the recovery file to be ignored.  Also, this behavior
could be abused in a FFA context by closing the app right before you get
killed.  Because you can leave and return without losing stats, this
could be a useful strategy for a certain kind of player.